+++ src/roboto/ai/__init__.py
@@ -4,10 +4,14 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from .record import PromptRequest
+from .record import (
+    PromptRequest,
+    SetSummaryRequest,
+)
 from .summary import AISummary
 
 __all__ = [
     "AISummary",
     "PromptRequest",
+    "SetSummaryRequest",
 ]